Move 'Tools' folder into new Tools namespace
