+++ Assets/Scripts/UI/Components/Specialised/Colour Picker/GlobalColourPicker.cs
@@ -1,5 +1,4 @@
 using PAC.UI.Components.General;
-using PAC.Drawing;
 using PAC.Input;
 using PAC.Maths;
 using PAC.Config;
@@ -7,6 +6,7 @@ using PAC.Config;
 using UnityEngine;
 using UnityEngine.Events;
 using PAC.Managers;
+using PAC.Tools;
 
 namespace PAC.UI.Components.Specialised.ColourPicker
 {
