+++ Assets/Tests/Drawing/CoordSnapping_Tests.cs
@@ -2,11 +2,11 @@ using System;
 
 using NUnit.Framework;
 
-using PAC.Drawing;
 using PAC.Extensions.System;
 using PAC.Geometry;
 using PAC.Geometry.Axes;
 using PAC.Geometry.Shapes;
+using PAC.Tools;
 
 namespace PAC.Tests.Drawing
 {
